+++ src/autocomplete-ids.ts
@@ -2,6 +2,7 @@
 let index = 0;
 export default class {
     ELEMENT: string;
+    LABEL: string;
     PREFIX: string;
     LIST: string;
     INPUT: string;
@@ -18,6 +19,7 @@ export default class {
         const id = optionId || elementId || '';
         this.PREFIX = `${id}aria-autocomplete-${index}`;
 
+        this.LABEL = `${this.PREFIX}-label`;
         this.LIST = `${this.PREFIX}-list`;
         this.BUTTON = `${this.PREFIX}-button`;
         this.OPTION = `${this.PREFIX}-option`;